Start multiplatform support
Just runtime for now. Tests can only run on JVM and Android due to use of JVM-specific coroutines 1.5.x test APIs.
