+++ .github/workflows/build.yaml
@@ -36,8 +36,8 @@ jobs:
       - run: ./gradlew -p molecule publish
         if: ${{ github.ref == 'refs/heads/trunk' && github.repository == 'cashapp/molecule' }}
         env:
-          ORG_GRADLE_PROJECT_SONATYPE_NEXUS_USERNAME: ${{ secrets.SONATYPE_NEXUS_USERNAME }}
-          ORG_GRADLE_PROJECT_SONATYPE_NEXUS_PASSWORD: ${{ secrets.SONATYPE_NEXUS_PASSWORD }}
+          ORG_GRADLE_PROJECT_mavenCentralUsername: ${{ secrets.SONATYPE_NEXUS_USERNAME }}
+          ORG_GRADLE_PROJECT_mavenCentralPassword: ${{ secrets.SONATYPE_NEXUS_PASSWORD }}
 
       - name: Deploy docs to website
         if: ${{ github.ref == 'refs/heads/trunk' }}
