+++ molecule/build.gradle
@@ -4,8 +4,8 @@ buildscript {
   dependencies {
     classpath deps.androidPlugin
     classpath deps.kotlin.gradlePlugin
-    classpath 'com.vanniktech:gradle-maven-publish-plugin:0.14.2'
-    classpath 'org.jetbrains.dokka:dokka-gradle-plugin:1.4.32'
+    classpath 'com.vanniktech:gradle-maven-publish-plugin:0.18.0'
+    classpath 'org.jetbrains.dokka:dokka-gradle-plugin:1.6.10'
     classpath deps.spotlessPlugin
   }
 
@@ -77,12 +77,11 @@ subprojects {
   }
 
   plugins.withId('maven-publish') {
-    mavenPublish {
-      targets {
-        inRepo {
-          def url = file("${rootProject.buildDir}/localMaven").toURI().toString()
-          releaseRepositoryUrl = url
-          snapshotRepositoryUrl = url
+    publishing {
+      repositories {
+        maven {
+          name = "installLocally"
+          url = "${rootProject.buildDir}/localMaven"
         }
       }
     }
