+++ molecule/gradle.properties
@@ -28,4 +28,8 @@ android.defaults.buildfeatures.renderscript=false
 android.defaults.buildfeatures.resvalues=false
 android.defaults.buildfeatures.shaders=false
 
+kotlin.js.compiler=ir
+kotlin.mpp.stability.nowarn=true
+kotlin.mpp.enableGranularSourceSetsMetadata=true
+
 systemProp.org.gradle.internal.http.socketTimeout=120000
