+++ molecule/molecule-runtime/src/commonJvmTest/kotlin/app/cash/molecule/MoleculeTest.kt
@@ -16,6 +16,7 @@
 package app.cash.molecule
 
 import androidx.compose.runtime.BroadcastFrameClock
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -28,20 +29,22 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.test.TestCoroutineDispatcher
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertSame
-import org.junit.Test
 import kotlin.coroutines.CoroutineContext
+import kotlin.test.Test
+import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
+import kotlin.test.assertSame
+import kotlin.test.fail
 
 @ExperimentalCoroutinesApi
-class MoleculeStateFlowTest {
+class MoleculeTest {
   @Test fun items() {
     val dispatcher = TestCoroutineDispatcher()
     val clock = BroadcastFrameClock()
     val scope = CoroutineScope(dispatcher + clock)
+    var value: Int? = null
 
-    val flow = scope.launchMolecule {
+    scope.launchMolecule(emitter = { value = it }) {
       var count by remember { mutableStateOf(0) }
       LaunchedEffect(Unit) {
         while (true) {
@@ -53,22 +56,22 @@ class MoleculeStateFlowTest {
       count
     }
 
-    assertEquals(0, flow.value)
+    assertEquals(0, value)
 
     clock.sendFrame(0)
-    assertEquals(0, flow.value)
+    assertEquals(0, value)
 
     dispatcher.advanceTimeBy(99)
     clock.sendFrame(0)
-    assertEquals(0, flow.value)
+    assertEquals(0, value)
 
     dispatcher.advanceTimeBy(1)
     clock.sendFrame(0)
-    assertEquals(1, flow.value)
+    assertEquals(1, value)
 
     dispatcher.advanceTimeBy(100)
     clock.sendFrame(0)
-    assertEquals(2, flow.value)
+    assertEquals(2, value)
 
     scope.cancel()
   }
@@ -80,7 +83,7 @@ class MoleculeStateFlowTest {
     // Use a custom subtype to prevent coroutines from breaking referential equality.
     val runtimeException = object : RuntimeException() {}
     val t = assertFailsWith<RuntimeException> {
-      scope.launchMolecule {
+      scope.launchMolecule(emitter = { fail() }) {
         throw runtimeException
       }
     }
@@ -103,18 +106,19 @@ class MoleculeStateFlowTest {
     val clock = BroadcastFrameClock()
     val exceptionHandler = RecordingExceptionHandler()
     val scope = CoroutineScope(dispatcher + clock + exceptionHandler)
+    var value: Int? = null
 
     // Use a custom subtype to prevent coroutines from breaking referential equality.
     val runtimeException = object : RuntimeException() {}
     var count by mutableStateOf(0)
-    val flow = scope.launchMolecule {
+    scope.launchMolecule(emitter = { value = it }) {
       if (count == 1) {
         throw runtimeException
       }
       count
     }
 
-    assertEquals(0, flow.value)
+    assertEquals(0, value)
 
     count++
     Snapshot.sendApplyNotifications() // Ensure external state mutation is observed.
@@ -129,10 +133,11 @@ class MoleculeStateFlowTest {
     val clock = BroadcastFrameClock()
     val exceptionHandler = RecordingExceptionHandler()
     val scope = CoroutineScope(dispatcher + clock + exceptionHandler)
+    var value: Int? = null
 
     // Use a custom subtype to prevent coroutines from breaking referential equality.
     val runtimeException = object : RuntimeException() {}
-    val flow = scope.launchMolecule {
+    scope.launchMolecule(emitter = { value = it }) {
       LaunchedEffect(Unit) {
         delay(50)
         throw runtimeException
@@ -140,12 +145,85 @@ class MoleculeStateFlowTest {
       0
     }
 
-    assertEquals(0, flow.value)
+    assertEquals(0, value)
 
     dispatcher.advanceTimeBy(50)
     clock.sendFrame(0)
     assertSame(runtimeException, exceptionHandler.exceptions.single())
 
     scope.cancel()
+  }
+
+  @Test fun errorInEmitterImmediately() {
+    val clock = BroadcastFrameClock()
+    val scope = CoroutineScope(clock)
+
+    // Use a custom subtype to prevent coroutines from breaking referential equality.
+    val runtimeException = object : RuntimeException() {}
+    val t = assertFailsWith<RuntimeException> {
+      scope.launchMolecule(emitter = { throw runtimeException }) {
+        0
+      }
+    }
+
+    assertSame(runtimeException, t)
+
+    scope.cancel()
+  }
+
+  @Test fun errorInEmitterDelayed() {
+    val dispatcher = TestCoroutineDispatcher()
+    val clock = BroadcastFrameClock()
+    val exceptionHandler = RecordingExceptionHandler()
+    val scope = CoroutineScope(dispatcher + clock + exceptionHandler)
+    var value: Int? = null
+
+    // Use a custom subtype to prevent coroutines from breaking referential equality.
+    val runtimeException = object : RuntimeException() {}
+    var count by mutableStateOf(0)
+    scope.launchMolecule(
+      emitter = {
+        if (it == 1) {
+          throw runtimeException
+        }
+        value = it
+      }
+    ) {
+      count
+    }
+
+    assertEquals(0, value)
+
+    count++
+    Snapshot.sendApplyNotifications() // Ensure external state mutation is observed.
+    clock.sendFrame(0)
+    assertSame(runtimeException, exceptionHandler.exceptions.single())
+
+    scope.cancel()
+  }
+
+  enum class DisposableEffectState { NOT_LAUNCHED, LAUNCHED, DISPOSED }
+
+  @Test fun disposableEffectDisposesWhenScopeIsCancelled() {
+    val dispatcher = TestCoroutineDispatcher()
+    val clock = BroadcastFrameClock()
+    val scope = CoroutineScope(dispatcher + clock)
+
+    var state: DisposableEffectState = DisposableEffectState.NOT_LAUNCHED
+
+    scope.launchMolecule {
+      DisposableEffect(Unit) {
+        state = DisposableEffectState.LAUNCHED
+
+        onDispose {
+          state = DisposableEffectState.DISPOSED
+        }
+      }
+    }
+
+    assertEquals(DisposableEffectState.LAUNCHED, state)
+
+    scope.cancel()
+    assertEquals(DisposableEffectState.DISPOSED, state)
   }
 }
